+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Cache-Control option when uploading files
 * Add `--lifecycleRule` to `create-bucket` and `update-bucket` and deprecate `--lifecycleRules` argument
 * Add extra dependencies for better UX, installable with `pip install b2[full]`
+* Add s3 endpoint to `get-account-info`
 
 ### Deprecated
 * Support of `-` as a valid filename in `upload-file` command. In future `-` will be an alias for standard input.